+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.0.9] - 2025-08-04
+
+- export of blocks and types with 'preservePath' option
+- new tools
+- some infos with attributes
+
 ## [0.0.8] - 2025-08-01
 
 - improved jsonrpc responses